[Description] Add screenshot showing the anatomy of a commit message
[Body] The first line of a commit message is its description. Many tools
use this line as a shorthand way of describing a commit. It is the first
impression of your commit, and should be less than 80 characters long.

All line following the description are the body of the commit. Use the
commit body to explain the reasoning behind your commit. Wrap the body
at 80 characters so that it more neatly prints to default sized
terminals.